LG-11116 Refactor Report Building (#9299)
* LG-11116 Refactor Report Building

changelog: Internal, Reporting, Reorganize Individual Metric Reports
