+++ app/jobs/reports/monthly_key_metrics_report.rb
@@ -6,9 +6,30 @@ module Reports
 
     attr_reader :report_date
 
-    def perform(date)
+    def perform(date = Time.zone.today)
       @report_date = date
-      csv_for_email = monthly_key_metrics_report_array
+
+      account_reuse_table = account_reuse_report.account_reuse_report
+      total_profiles_table = account_reuse_report.total_identities_report
+
+      upload_to_s3(account_reuse_table, report_name: 'account_reuse')
+      upload_to_s3(total_profiles_table, report_name: 'total_profiles')
+
+      email_tables = [
+        [
+          {
+            title: "IDV app reuse rate #{account_reuse_report.stats_month}",
+            float_as_percent: true,
+            precision: 4,
+          },
+          *account_reuse_table,
+        ],
+        [
+          { title: 'Total proofed identities' },
+          *total_profiles_table,
+        ],
+      ]
+
       email_message = "Report: #{REPORT_NAME} #{date}"
       email_addresses = emails.select(&:present?)
 
@@ -17,7 +38,7 @@ module Reports
           email: email_addresses,
           subject: "Monthly Key Metrics Report - #{date}",
           message: email_message,
-          tables: csv_for_email,
+          tables: email_tables,
         ).deliver_now
       else
         Rails.logger.warn 'No email addresses received - Monthly Key Metrics Report NOT SENT'
@@ -32,19 +53,29 @@ module Reports
       emails
     end
 
-    def monthly_key_metrics_report_array
-      csv_array = []
-
-      account_reuse_report_csv.each do |row|
-        csv_array << row
-      end
-
-      csv_array
+    def account_reuse_report
+      @account_reuse_report ||= Reporting::AccountReuseAndTotalIdentitiesReport.new(report_date)
     end
 
-    # Individual Key Metric Report
-    def account_reuse_report_csv
-      Reports::MonthlyAccountReuseReport.new(report_date).report_csv
+    def upload_to_s3(report_body, report_name: nil)
+      _latest, path = generate_s3_paths(REPORT_NAME, 'csv', subname: report_name, now: report_date)
+
+      if bucket_name.present?
+        upload_file_to_s3_bucket(
+          path: path,
+          body: csv_file(report_body),
+          content_type: 'text/csv',
+          bucket: bucket_name,
+        )
+      end
+    end
+
+    def csv_file(report_array)
+      CSV.generate do |csv|
+        report_array.each do |row|
+          csv << row
+        end
+      end
     end
   end
 end
