+++ spec/jobs/reports/monthly_key_metrics_report_spec.rb
@@ -7,12 +7,31 @@ RSpec.describe Reports::MonthlyKeyMetricsReport do
   let(:name) { 'monthly-key-metrics-report' }
   let(:agnes_email) { 'fake@agnes_email.com' }
   let(:feds_email) { 'fake@feds_email.com' }
+  let(:s3_report_bucket_prefix) { 'reports-bucket' }
+  let(:account_reuse_s3_path) do
+    'int/monthly-key-metrics-report/2021/2021-03-02.monthly-key-metrics-report/account_reuse.csv'
+  end
+  let(:total_profiles_s3_path) do
+    'int/monthly-key-metrics-report/2021/2021-03-02.monthly-key-metrics-report/total_profiles.csv'
+  end
 
   before do
     allow(IdentityConfig.store).to receive(:team_agnes_email).
       and_return(agnes_email)
     allow(IdentityConfig.store).to receive(:team_all_feds_email).
       and_return(feds_email)
+
+    allow(Identity::Hostdata).to receive(:env).and_return('int')
+    allow(Identity::Hostdata).to receive(:aws_account_id).and_return('1234')
+    allow(Identity::Hostdata).to receive(:aws_region).and_return('us-west-1')
+    allow(IdentityConfig.store).to receive(:s3_report_bucket_prefix).
+      and_return(s3_report_bucket_prefix)
+
+    Aws.config[:s3] = {
+      stub_responses: {
+        put_object: {},
+      },
+    }
   end
 
   it 'sends out a report to the email listed with one total user' do
@@ -48,6 +67,24 @@ RSpec.describe Reports::MonthlyKeyMetricsReport do
       subject: 'Monthly Key Metrics Report - 2021-03-02',
       tables: anything,
     ).and_call_original
+
+    subject.perform(report_date)
+  end
+
+  it 'uploads a file to S3 based on the report date' do
+    expect(subject).to receive(:upload_file_to_s3_bucket).with(
+      path: account_reuse_s3_path,
+      body: anything,
+      content_type: 'text/csv',
+      bucket: 'reports-bucket.1234-us-west-1',
+    ).exactly(1).time.and_call_original
+
+    expect(subject).to receive(:upload_file_to_s3_bucket).with(
+      path: total_profiles_s3_path,
+      body: anything,
+      content_type: 'text/csv',
+      bucket: 'reports-bucket.1234-us-west-1',
+    ).exactly(1).time.and_call_original
 
     subject.perform(report_date)
   end
